+++ post-processor/docker-push/post-processor_test.go
@@ -2,7 +2,9 @@ package dockerpush
 
 import (
 	"bytes"
+	"github.com/mitchellh/packer/builder/docker"
 	"github.com/mitchellh/packer/packer"
+	"github.com/mitchellh/packer/post-processor/docker-import"
 	"testing"
 )
 
@@ -28,4 +30,85 @@ func testUi() *packer.BasicUi {
 
 func TestPostProcessor_ImplementsPostProcessor(t *testing.T) {
 	var _ packer.PostProcessor = new(PostProcessor)
+}
+
+func TestPostProcessor_PostProcess(t *testing.T) {
+	driver := &docker.MockDriver{}
+	p := &PostProcessor{Driver: driver}
+	artifact := &packer.MockArtifact{
+		BuilderIdValue: dockerimport.BuilderId,
+		IdValue: "foo/bar",
+	}
+
+	result, keep, err := p.PostProcess(testUi(), artifact)
+	if result != nil {
+		t.Fatal("should be nil")
+	}
+	if keep {
+		t.Fatal("should not keep")
+	}
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if !driver.PushCalled {
+		t.Fatal("should call push")
+	}
+	if driver.PushName != "foo/bar" {
+		t.Fatal("bad name")
+	}
+}
+
+func TestPostProcessor_PostProcess_portInName(t *testing.T) {
+	driver := &docker.MockDriver{}
+	p := &PostProcessor{Driver: driver}
+	artifact := &packer.MockArtifact{
+		BuilderIdValue: dockerimport.BuilderId,
+		IdValue: "localhost:5000/foo/bar",
+	}
+
+	result, keep, err := p.PostProcess(testUi(), artifact)
+	if result != nil {
+		t.Fatal("should be nil")
+	}
+	if keep {
+		t.Fatal("should not keep")
+	}
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if !driver.PushCalled {
+		t.Fatal("should call push")
+	}
+	if driver.PushName != "localhost:5000/foo/bar" {
+		t.Fatal("bad name")
+	}
+}
+
+func TestPostProcessor_PostProcess_tags(t *testing.T) {
+	driver := &docker.MockDriver{}
+	p := &PostProcessor{Driver: driver}
+	artifact := &packer.MockArtifact{
+		BuilderIdValue: dockerimport.BuilderId,
+		IdValue: "hashicorp/ubuntu:precise",
+	}
+
+	result, keep, err := p.PostProcess(testUi(), artifact)
+	if result != nil {
+		t.Fatal("should be nil")
+	}
+	if keep {
+		t.Fatal("should not keep")
+	}
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if !driver.PushCalled {
+		t.Fatal("should call push")
+	}
+	if driver.PushName != "hashicorp/ubuntu" {
+		t.Fatalf("bad name: %s", driver.PushName)
+	}
 }